Refactor chat DTOs into dedicated classes and files
