+++ apps/api/src/modules/chat/dto/chat-message.dto.ts
@@ -1,0 +1,9 @@
+﻿import { IsString } from 'class-validator';
+
+export class ChatMessageDto {
+  @IsString()
+  role!: 'user' | 'assistant' | 'system';
+
+  @IsString()
+  content!: string;
+}
