+++ apps/api/src/modules/chat/dto/chat-response.dto.ts
@@ -1,0 +1,6 @@
+﻿import { IsString } from 'class-validator';
+
+export class ChatResponseDto {
+  @IsString()
+  content!: string;
+}
