+++ apps/api/src/modules/chat/dto/index.ts
@@ -1,0 +1,3 @@
+﻿export * from './chat-message.dto';
+export * from './chat-request.dto';
+export * from './chat-response.dto';
